fix: Crash calling DetailView occurring on Review
On the review, the app crash because the CREATOR variable of the Parcelable class (Movie.java) was not public. Following the advice of the reviewer the variable will be now public.
Note: I could not make the app crash with the same emulator (Nexus 6P API25) the reviewer used. I tested it before with others emulators and it worked without any issues. Strange!!